// Gerdau/web/src/redux/company/types.ts
export enum CompanyActionTypes {
  FETCH_PLANTS = "FETCH_PLANTS",
  FETCH_PLANTS_SUCCESS = "FETCH_PLANTS_SUCCESS",
  FETCH_PLANTS_ERROR = "FETCH_PLANTS_ERROR",

  FETCH_SUBLEVELS = "FETCH_MODELS",
  FETCH_SUBLEVELS_SUCCESS = "FETCH_MODELS_SUCCESS",
  FETCH_SUBLEVELS_ERROR = "FETCH_MODELS_ERROR",

  FETCH_AREAS = "FETCH_AREAS",
  FETCH_AREAS_SUCCESS = "FETCH_AREAS_SUCCESS",
  FETCH_AREAS_ERROR = "FETCH_AREAS_ERROR",
}